// web/snake/src/features/dashboard/types.ts
export interface TokenStat {
  token_id: number
  token_name: string
  quota: number
  count: number
  tokens: number
}